refactor FirebaseAuthContext
