+++ packages/Framework/src/Exports/Managers.ts
@@ -1,2 +1,2 @@
-export * from '../Managers/FirebaseAuthContext';
+export * from '../Managers/Firebase/FirebaseAuthContext';
 export * from '../Managers/LocalDataContext';
